swap values using "pointers"
This program demonstrates swapping the values of two integers using pointers. The swap function takes two integer pointers as arguments and exchanges their values by temporarily storing one value in a variable temp. In the main function, two integers (num1 and num2) are defined and displayed before swapping. The swap function is then called with the addresses of these integers, and the swapped values are displayed afterward. This implementation highlights the use of pointers to manipulate variables indirectly and perform value swapping.